Suppress warning about mismatched signed comparison

+++ src/utils.h
@@ -260,7 +260,7 @@ size_t find_first_line(
 
 inline bool
 matches(const char* start, const char* end, const std::string& needle) {
-  if (end <= start || end - start < needle.size()) {
+  if (end <= start || static_cast<unsigned long>(end - start) < needle.size()) {
     return false;
   }
   bool res = strncmp(start, needle.data(), needle.size()) == 0;
